- refactored to have seperate site and app install/deploy steps - this includes also having site and app deploy keys - added to a deploy user a small bit of code to ensure ssh-agent is loaded and running - this makes the environment the same for fabops site updates as helper script site updates

+++ fabops/nginx.py
@@ -12,18 +12,26 @@ from fabric.context_managers import cd
 
 import fabops.common
 
-_version  = '1.2.6'
-_tarball  = 'nginx-%s.tar.gz' % _version
-_url      = 'http://nginx.org/download/%s' % _tarball
-_tmp_dir  = '/tmp/nginx-%s' % _version
-_username = 'nginx'
-
+_version   = '1.4.0'
+_tarball   = 'nginx-%s.tar.gz' % _version
+_url       = 'http://nginx.org/download/%s' % _tarball
+_tmp_dir   = '/tmp/nginx-%s' % _version
+_username  = 'nginx'
+_configure = ' '.join(['--prefix=/usr/local/nginx',
+                       '--conf-path=/etc/nginx/nginx.conf',
+                       '--pid-path=/var/run/nginx.pid',
+                       '--user=%s' % _username,
+                       '--with-http_stub_status_module',
+                       '--with-http_ssl_module',
+                      ])
 @task
 def download():
     """
     Download and extract the nginx tarball
     """
     with cd('/tmp'):
+        if exists(_tarball):
+            run('rm -f %s' % _tarball)
         run('wget %s' % _url)
         run('tar xf %s' % _tarball)
 
@@ -34,7 +42,7 @@ def build():
     """
     if exists(_tmp_dir):
         with cd(_tmp_dir):
-            run('./configure --conf-path=/etc/nginx/nginx.conf --pid-path=/var/run/nginx.pid --user=nginx --with-http_stub_status_module')
+            run('./configure %s' % _configure)
             run('make')
 
 @task
@@ -49,7 +57,7 @@ def install(force=False):
     if not force and fabops.common.user_exists(_username):
         print('nginx user already exists, skipping nginx install')
     else:
-        for p in ('build-essential', 'libpcre3-dev', 'zlib1g-dev'):
+        for p in ('build-essential', 'libssl-dev', 'libpcre3-dev', 'zlib1g-dev'):
             fabops.common.install_package(p)
 
         download()
@@ -90,42 +98,38 @@ def install_site(siteName, siteConfig):
 
         if 'nginx.root' not in cfg:
             print('nginx.root not found in site configuration')
-        s = ''
-        if 'nginx.listen' in cfg:
-            s = cfg['nginx.listen']
-        if 'nginx.port' in cfg:
-            if len(s) > 0:
-                s += ':'
-            s += str(cfg['nginx.port'])
-        cfg['nginx.listen'] = s
 
         upload_template(os.path.join(siteConfig['site_config_dir'], '%s.nginx' % cfg['name']),
-                        '/etc/nginx/conf.d/%s' % cfg['name'],
+                        '/etc/nginx/conf.d/%s.conf' % cfg['name'],
                         context=cfg,
                         use_sudo=True)
-        sudo('chown root:root /etc/nginx/conf.d/%s' % cfg['name'])
-        sudo('mkdir -p %s' % cfg['nginx.root'])
-        sudo('chown nginx:nginx %s' % cfg['nginx.root'])
+        sudo('chown root:root /etc/nginx/conf.d/%s.conf' % cfg['name'])
+        if not exists(cfg['nginx.root']):
+            sudo('mkdir -p %s' % cfg['nginx.root'])
+            sudo('chown nginx:nginx %s' % cfg['nginx.root'])
 
 @task
 def deploy_site(siteConfig):
     """
     Deploy an installed nginx site
     """
-    if fabops.common.user_exists(siteConfig['deploy_user']):
-        with settings(user=siteConfig['deploy_user'], key_filename='/home/%s/.ssh/%s' % (siteConfig['deploy_user'], siteConfig['deploy_key'])):
-            tempDir = '/tmp/%s' % siteConfig['name']
-            workDir = os.path.join(tempDir, siteConfig['name'])
+    if 'repository' in siteConfig:
+        if fabops.common.user_exists(siteConfig['deploy_user']):
+            with settings(user=siteConfig['deploy_user'], key_filename='/home/%s/.ssh/%s' % (siteConfig['deploy_user'], siteConfig['deploy_key'])):
+                tempDir = '/tmp/%s' % siteConfig['name']
+                workDir = os.path.join(tempDir, siteConfig['name'])
 
-            if exists(workDir):
-                run('rm -rf %s' % workDir)
-            else:
-                run('mkdir -p %s' % workDir)
-            
-            run('git clone %s %s' % (siteConfig['repository']['url'], workDir))
+                if exists(workDir):
+                    run('rm -rf %s' % workDir)
+                else:
+                    run('mkdir -p %s' % workDir)
+                
+                run('git clone %s %s' % (siteConfig['repository']['url'], workDir))
 
-            with cd(workDir):
-                run('git checkout %s' % siteConfig['deploy_branch'])
-                sudo('cp %s/* %s' % (workDir, siteConfig['nginx']['root']))
-        if sudo('service nginx testconfig'):
-            sudo('service nginx restart')
+                with cd(workDir):
+                    run('git checkout %s' % siteConfig['deploy_branch'])
+                    sudo('cp %s/* %s' % (workDir, siteConfig['nginx']['root']))
+            if sudo('service nginx testconfig'):
+                sudo('service nginx restart')
+    else:
+        print('site does not have a defined repository, skipping')
